update walk model and routes

+++ routes/walk.routes.js
@@ -1,21 +1,27 @@
 const express = require('express');
 const router = express.Router();
 const Walk = require('../models/Walk.model');
+const { isAuthenticated } = require("../middleware/jwt.middleware.js");
 
 
-router.post('/', async (req, res, next) => {
+router.post('/', isAuthenticated, async (req, res, next) => {
+    const authenticatedUserId = req.payload._id;
     try {
-        const savedWalk = await Walk.create(req.body);
+        const walkData = { ...req.body, user: authenticatedUserId };
+        const savedWalk = await Walk.create(walkData);
         res.status(201).json(savedWalk);
     } catch (error) {
         next(error);
     }
 });
 
+
 router.get('/', async (req, res, next) => {
-    //todo: get only current user walks
     try {
-        const walks = await Walk.find();
+        const walks = await Walk.find().populate({
+            path: 'user',
+            select: 'name'
+        });
         res.json(walks);
     } catch (error) {
         next(error)
@@ -25,31 +31,54 @@ router.get('/', async (req, res, next) => {
 router.get('/:walkId', async (req, res, next) => {
     const { walkId } = req.params;
     try {
-        const walk = await Walk.findById(walkId);
+        const walk = await Walk.findById(walkId).populate({
+            path: 'user',
+            select: 'name'
+        });
         res.json(walk);
     } catch (error) {
         next(error)
     }
 });
 
-router.put('/:walkId', async (req, res, next) => {
+router.put('/:walkId', isAuthenticated, async (req, res, next) => {
     const { walkId } = req.params;
+    const authenticatedUserId = req.payload._id;
+
     try {
-        const updatedWalk = await Walk.findByIdAndUpdate(walkId, req.body, { new: true });
+        const walk = await Walk.findById(walkId).populate('user', 'name');
+        if (walk.user._id.toString() !== authenticatedUserId.toString()) {
+            return res.status(403).json({ message: 'You are not authorized to update this walk' });
+        }
+
+        const updatedWalk = await Walk.findByIdAndUpdate(walkId, req.body, { new: true }).populate({
+            path: 'user',
+            select: 'name'
+        });
+        
         res.json(updatedWalk);
     } catch (error) {
-        next(error)
+        next(error);
     }
 });
 
-router.delete('/:walkId', async (req, res, next) => {
+
+router.delete('/:walkId', isAuthenticated, async (req, res, next) => {
     const { walkId } = req.params;
+    const authenticatedUserId = req.payload._id;
+
     try {
+        const walk = await Walk.findById(walkId).populate('user', 'name');
+        if (walk.user._id.toString() !== authenticatedUserId.toString()) {
+            return res.status(403).json({ message: 'You are not authorized to delete this walk' });
+        }
+
         await Walk.findByIdAndDelete(walkId);
         res.status(204).send();
     } catch (error) {
-        next(error)
+        next(error);
     }
 });
+
 
 module.exports = router;
